[#38] refactor: MainController에서 메소드 분리 및 Common 패키지 생성
- MainController에서 Reservation기능을 다른 컨트롤러로 분리
- 공통으로 들어가야할 경우를 대비해서 Common 패키지 생성
