+++ BE/src/main/java/com/codesquad/airbnb/reservation/domain/Guest.java
@@ -5,8 +5,6 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-import java.time.LocalDate;
-
 @Getter
 @NoArgsConstructor
 @Setter
